pyproject.toml
[tool.poetry]
authors = ["mm21 <mm21.dev@gmail.com>"]
classifiers = [
  # Get the list of trove classifiers here: https://pypi.org/classifiers/
  "Programming Language :: Python :: Implementation :: CPython",
  "Operating System :: OS Independent",
  "Typing :: Typed",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Development Status :: 2 - Pre-Alpha",
  "Natural Language :: English",
  "License :: OSI Approved :: MIT License"
]
description = "Hydrate your files to build a database with rich interconnectivity between your data"
homepage = "https://github.com/mm21/protium-db"
name = "protium-db"
packages = [{include = "protium"}]
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
python = "^3.12"
pyrollup = "^0.1.0"
pydantic = "^2.8.2"
pyyaml = "^6.0.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
black = "^24.8.0"

[tool.black]
include = '\.pyi?$|SConstruct$'
line-length = 80

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]
